requirements.txt
langchain
langchain-community
openai
tiktoken
faiss-cpu
pydantic
fastapi
uvicorn
python-dotenv
pypdf
pymupdf